Initial implementation of performance schema.

+++ types/kotlin2/Screen.kt
@@ -25,7 +25,10 @@ import im.vector.app.features.analytics.itf.VectorAnalyticsScreen
  * Triggered when the user changed screen
  */
 data class Screen(
-    val durationMs: Double? = null,
+    /**
+     * How long the screen was displayed for in milliseconds.
+     */
+    val durationMs: Int? = null,
     val screenName: ScreenName,
 ) : VectorAnalyticsScreen {
